Moved hint events into separate js files
