+++ js/modals/question-modal/events/hints/save-hint.js
@@ -1,0 +1,8 @@
+
+function saveHint(hintText) {
+   'use strict';
+
+   $('.hint.active').data('hint-text', hintText);
+}
+
+module.exports = saveHint;
